securing the views

+++ products/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, get_object_or_404, redirect, reverse
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.db.models.functions import Lower
 
@@ -72,8 +73,16 @@ def product_detail(request, product_id):
     return render(request, 'products/product_detail.html', context)
 
 
+@login_required
 def add_product(request):
     """ Add Product to the store"""
+
+    if not request.user.is_superuser():
+        messages.error(
+            request,
+            'Sorry, only store owners can do that!'
+        )
+        return redirect(reverse('home'))
 
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
@@ -100,8 +109,17 @@ def add_product(request):
     return render(request, template, context)
 
 
+@login_required
 def edit_product(request, product_id):
     """ Edit product in the store """
+    
+    if not request.user.is_superuser():
+        messages.error(
+            request,
+            'Sorry, only store owners can do that!'
+        )
+        return redirect(reverse('home'))
+    
     product = get_object_or_404(Product, pk=product_id)
 
     if request.method == 'POST':
@@ -134,10 +152,19 @@ def edit_product(request, product_id):
     return render(request, template, context)
 
 
+@login_required
 def delete_product(request, product_id):
     """ delete product in the store """
-    product = get_object_or_404(Product, pk=product_id)
     
+    if not request.user.is_superuser():
+        messages.error(
+            request,
+            'Sorry, only store owners can do that!'
+        )
+        return redirect(reverse('home'))
+    
+    product = get_object_or_404(Product, pk=product_id)
+
     product.delete()
     messages.success(
         request,
